Split unit test by MVT pattern
