+++ linku/moim/tests/test_views.py
@@ -3,40 +3,6 @@ import datetime
 from moim.models import Meeting, Applier
 
 
-def test_meeting_model_have_fields():
-    Meeting._meta.get_field('maker')
-    Meeting._meta.get_field('name')
-    Meeting._meta.get_field('place')
-    Meeting._meta.get_field('start_time')
-    Meeting._meta.get_field('image_path')
-    Meeting._meta.get_field('distance_near_univ')
-    Meeting._meta.get_field('price_range')
-
-
-def test_applier_model_have_fields():
-    Applier._meta.get_field('name')
-    Applier._meta.get_field('phone_number')
-    Applier._meta.get_field('gender')
-    Applier._meta.get_field('meeting')
-
-
-@pytest.mark.django_db
-def test_homepage_use_meeting_list_template(client):
-    response_templates = client.get('/').templates
-    assert 'meeting_list.html' in (
-        template.name for template in response_templates)
-
-
-@pytest.mark.django_db
-def test_apply_meeting_view_use_correct_template(client):
-    Meeting.objects.create(maker='test maker', name='test name', place='test place',
-                           start_time=datetime.datetime.now(),
-                           distance_near_univ='test distance_near_univ', price_range='test price_range')
-    response_templates = client.get('/meetings/1/apply/').templates
-    assert 'apply_meeting.html' in (
-        template.name for template in response_templates)
-
-
 @pytest.mark.django_db
 def test_apply_meeting_view_return_correct_meeting_name(client):
     Meeting.objects.create(maker='test maker1', name='test name1', place='test place1',
@@ -48,14 +14,6 @@ def test_apply_meeting_view_return_correct_meeting_name(client):
     meeting = Meeting.objects.get(maker='test maker2')
     response = client.get('/meetings/' + str(meeting.id) + '/apply/')
     assert meeting.name in response.content.decode("utf8")
-
-
-@pytest.mark.django_db
-def test_save_meeting():
-    Meeting.objects.create(maker='test maker', name='test name', place='test place',
-                           start_time=datetime.datetime.now(),
-                           distance_near_univ='test distance_near_univ', price_range='test price_range')
-    Meeting.objects.get(name='test name')
 
 
 @pytest.mark.django_db
@@ -90,13 +48,6 @@ def test_homepage_view_multiple_cards(client):
     assert "test name1" in response.content.decode("utf8")
     assert "test name2" in response.content.decode("utf8")
 
-
-@pytest.mark.django_db
-def test_save_applier():
-    start_time = datetime.datetime.now()
-    meeting = Meeting.objects.create(maker='test maker', name='test name1', place='test place', start_time=start_time,
-                                     distance_near_univ='test distance_near_univ', price_range='test price_range')
-    Applier.objects.create(name='test name', phone_number='010-1111-1111', gender='M', meeting=meeting)
 
 
 @pytest.mark.django_db
